# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RFA knowledge server — /ask
  summary: desk(대응 에이전트) ↔ 지식 서버 계약
  description: 'head → broker → task agent → censor 를 수행하는 `ask()` 하나를 노출한다. `audience` 가 검열 프로파일을 정한다(public→public,
    company/self→internal). 동기 응답(서버 타임아웃 180초 → refusal no_knowledge/timeout). 같은 request_id 는 캐시 응답.
    `feedback[].reason` 은 사람 입력으로 신뢰되어 검열 규칙(learned.yaml)으로 되먹임된다. 인증: `Authorization: Bearer <RFA_ASK_TOKEN>`
    (.env.dev).'
  version: 1.0.0
paths:
  /ask:
    post:
      tags:
      - ask
      summary: Post Ask
      operationId: ask
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AskRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AskResponse'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Unauthorized
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Service Unavailable
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /chat/sync:
    post:
      tags:
      - personal
      summary: Post Chat Sync
      description: 개인 채팅 JSON 동기 버전(CLI 용, audience self). 프런트는 SSE `POST /chat` 을 쓴다.
      operationId: chatSync
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AskResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /teams:
    post:
      tags:
      - teams
      summary: Post Team
      description: 요구사항(자연어 name/description) → 승인된 역할 카탈로그(roles.yaml) 안에서 패터닝 → 기본 샌드박스에 supervisor(secondary)
        + 멤버를 선언(teams.yaml)·적용(agents apply)·시드하고 /ask 카탈로그에 등록한다.
      operationId: createTeam
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '200':
          description: 같은 task_id 의 기존 팀 (멱등)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Unprocessable Entity
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Unauthorized
    get:
      tags:
      - teams
      summary: List Teams
      operationId: listTeams
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /teams/{team_id}:
    get:
      tags:
      - teams
      summary: Get Team
      operationId: getTeam
      parameters:
      - name: team_id
        in: path
        required: true
        schema:
          type: string
          title: Team Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - teams
      summary: Delete Team
      operationId: removeTeam
      parameters:
      - name: team_id
        in: path
        required: true
        schema:
          type: string
          title: Team Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /chat:
    post:
      tags:
      - frontend
      summary: Chat Stream
      description: '검색하거나 물어보기 SSE. 담당자 탐색(점수 ≥ head.select_threshold 전부 병렬 위임) → 위임 → 합치기/직접 답변 → 등급
        검사(owner 사내, guest 사외) → 답변. 봉투 `{type, runId, seq, ts, data}`; type: `run.start` {runId, conversationId,
        messageId, role, agentId, level: public|company, requestId}, `assistant.delta` {text} — the 요청
        파악 preface, `agents.search` {query, candidates: [{agentId, reason, score 0..1}]}, `agents.select`
        {selected: [{agentId, task}], skipped: [{agentId, reason}]} — every candidate ≥ head.select_threshold,
        `delegate.start` {callId, agentId, task}, `delegate.log` {callId, text} — guest: fixed phrases
        only, `delegate.end` {callId, status: ok|none|blocked|error, summary, refs: [], durationMs}, `guard.start`
        {level: public(사외)|company(사내)}, `guard.end` {status: pass|redacted|blocked, note, redactions:
        int}, `answer.delta` {text} — one sentence of the checked answer, `run.end` {status: done|error,
        durationMs, messageId, error?, code?} (stop = the client closes the stream)'
      operationId: chatStream
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatStreamRequest'
      responses:
        '200':
          description: SSE stream of SseEnvelope frames
          content:
            application/json:
              schema: {}
            text/event-stream:
              schema:
                description: 'Every SSE event of every stream: ``{type, runId, seq, ts, data}``. ``type``
                  names may change

                  with the front-end; the envelope does not.'
                properties:
                  type:
                    description: event name, e.g. run.started | stage | delta | guard.final | done | error
                    title: Type
                    type: string
                  runId:
                    title: Runid
                    type: string
                  seq:
                    description: 0-based, strictly increasing within a run
                    minimum: 0
                    title: Seq
                    type: integer
                  ts:
                    description: unix seconds
                    title: Ts
                    type: number
                  data:
                    additionalProperties: true
                    title: Data
                    type: object
                required:
                - type
                - runId
                - seq
                - ts
                title: SseEnvelope
                type: object
        '409':
          description: a run is already streaming in this conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/head/ask:
    post:
      tags:
      - head
      summary: Head Ask
      description: 'RFA_module head 계약 v0.2.0 호환(D-21): RFA_module `.env` 의 `HEAD_URL=http://127.0.0.1:8799/v1/head`.
        인증 없이 루프백만 받는다. request_id 는 url+차수(거절 수+1)로 서버가 만들어 재시도에 멱등, refusal 은 한국어 문자열, 크기 초과는 잘라서 처리,
        서버 한도 110초. 들어오는 즉시 결재함에 「작성 중」 항목이 생긴다.'
      operationId: headAsk
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HeadAskIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HeadAskOut'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    AskRequest:
      properties:
        request_id:
          type: string
          pattern: ^[A-Za-z0-9_.:-]{1,64}$
          title: Request Id
          description: 멱등키. 같은 id 재요청 시 캐시 응답
        question:
          type: string
          maxLength: 10000
          minLength: 1
          title: Question
        channel:
          type: string
          enum:
          - github
          - slack
          - cli
          - web
          title: Channel
          description: github|slack (desk). cli|web 은 개인 채팅
        audience:
          type: string
          enum:
          - public
          - company
          - self
          title: Audience
          description: public→censor profile public, company/self→internal
        target:
          type: string
          maxLength: 500
          title: Target
          description: 스레드·이슈 식별자(계보 추적)
          default: ''
        url:
          anyOf:
          - type: string
            maxLength: 2000
          - type: 'null'
          title: Url
        requester:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Requester
        context:
          items:
            $ref: '#/components/schemas/ContextItem'
          type: array
          maxItems: 50
          title: Context
        feedback:
          items:
            $ref: '#/components/schemas/FeedbackItem'
          type: array
          maxItems: 20
          title: Feedback
      type: object
      required:
      - request_id
      - question
      - channel
      - audience
      title: AskRequest
    AskResponse:
      properties:
        request_id:
          type: string
          title: Request Id
        knowledge:
          type: string
          title: Knowledge
          description: 검열을 통과한 텍스트. refusal 이 있으면 빈 문자열
        task:
          anyOf:
          - $ref: '#/components/schemas/TaskRef'
          - type: 'null'
        refusal:
          anyOf:
          - $ref: '#/components/schemas/Refusal'
          - type: 'null'
        censor:
          $ref: '#/components/schemas/CensorSummary'
      type: object
      required:
      - request_id
      - knowledge
      - task
      - refusal
      - censor
      title: AskResponse
...
    CensorSummary:
      properties:
        profile:
          type: string
          title: Profile
        verdict:
          type: string
          enum:
          - allow
          - redact
          title: Verdict
        redactions:
          items:
            $ref: '#/components/schemas/Redaction'
          type: array
          title: Redactions
      type: object
      required:
      - profile
      - verdict
      title: CensorSummary
...
    ChatRequest:
      properties:
        question:
          type: string
          maxLength: 10000
          minLength: 1
          title: Question
        session_id:
          anyOf:
          - type: string
            pattern: ^[A-Za-z0-9_.:-]{1,64}$
          - type: 'null'
          title: Session Id
        channel:
          type: string
          enum:
          - cli
          - web
          title: Channel
          default: web
      type: object
      required:
      - question
      title: ChatRequest
      description: Personal chat (audience self) — same ``ask()`` function, request_id issued by the server.
    ChatStreamRequest:
      properties:
        text:
          type: string
          maxLength: 10000
          minLength: 1
          title: Text
        role:
          type: string
          enum:
          - owner
          - guest
          title: Role
          default: guest
        conversationId:
          anyOf:
          - type: string
            maxLength: 64
            pattern: ^[A-Za-z0-9_.:-]{1,64}$
          - type: 'null'
          title: Conversationid
        agentId:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Agentid
          description: 'pinned chat partner (`#chat/{agentId}`): skip ranking, delegate to it'
        history:
          items:
            $ref: '#/components/schemas/HistoryItem'
          type: array
          maxItems: 50
          title: History
          description: guest only
      type: object
      required:
      - text
      title: ChatStreamRequest
      description: '``POST /chat`` (SSE). ``role`` is what the front-end claims; only an authenticated
        caller

        (bearer RFA_ASK_TOKEN) is granted ``owner`` — anyone else is treated as ``guest``. The owner''s

        conversation is stored (earlier turns come from the server); a guest''s is not, so a guest sends
        its

        earlier turns in ``history``.'
...
    ContextItem:
      properties:
        author:
          type: string
          maxLength: 200
          title: Author
          default: ''
        text:
          type: string
          maxLength: 8000
          title: Text
        at:
          type: string
          maxLength: 64
          title: At
          default: ''
      type: object
      required:
      - text
      title: ContextItem
      description: 'One message of the external thread (untrusted input: shown to the head as data only).'
...
    ErrorBody:
      properties:
        code:
          type: string
          title: Code
        detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Detail
      type: object
      required:
      - code
      title: ErrorBody
    ErrorOut:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
      type: object
      required:
      - code
      - message
      title: ErrorOut
      description: 'Front-end error body: ``code`` for the client, ``message`` ready to show (Korean).'
    FeedbackItem:
      properties:
        draft:
          type: string
          maxLength: 20000
          title: Draft
          default: ''
        reason:
          type: string
          maxLength: 1000
          minLength: 1
          title: Reason
        at:
          type: string
          maxLength: 64
          title: At
          default: ''
      type: object
      required:
      - reason
      title: FeedbackItem
      description: 'One rejected draft. ``reason`` is human input (trusted, fed back into the censor rules);

        ``draft`` is external text (data only).'
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HeadAskIn:
      properties:
        question:
          type: string
          title: Question
        channel:
          type: string
          enum:
          - github
          - slack
          title: Channel
        audience:
          type: string
          enum:
          - public
          - company
          title: Audience
        target:
          type: string
          title: Target
        url:
          type: string
          title: Url
        requester:
          type: string
          title: Requester
          default: ''
        context:
          items:
            $ref: '#/components/schemas/ThreadMessageIn'
          type: array
          title: Context
        feedback:
          items:
            $ref: '#/components/schemas/RejectionIn'
          type: array
          title: Feedback
          description: every earlier rejection of this approval
      type: object
      required:
      - question
      - channel
      - audience
      - target
      - url
      title: HeadAskIn
      description: 'RFA_module ``AskRequest``. No size limits here: oversized fields are cut, never refused.'
    HeadAskOut:
      properties:
        knowledge:
          type: string
          title: Knowledge
        task:
          anyOf:
          - $ref: '#/components/schemas/TaskRefOut'
          - type: 'null'
        refusal:
          anyOf:
          - type: string
          - type: 'null'
          title: Refusal
        requestId:
          type: string
          title: Requestid
          description: 'derived: rfa-<sha1(url)>-r<round>; retries of a round share it'
        itemId:
          type: string
          title: Itemid
          description: 결재함 item id of this mention
        round:
          type: integer
          title: Round
        grade:
          type: string
          enum:
          - public
          - company
          title: Grade
          description: the request's grade (source grade wins, D-24)
        gradeSource:
          type: string
          enum:
          - source
          - channel
          title: Gradesource
        sourceId:
          anyOf:
          - type: string
          - type: 'null'
          title: Sourceid
        censor:
          additionalProperties: true
          type: object
          title: Censor
          default: {}
      type: object
      required:
      - knowledge
      - task
      - refusal
      - requestId
      - itemId
      - round
      - grade
      - gradeSource
      title: HeadAskOut
      description: 'RFA_module ``AskResponse`` (+ rfa_mas extras). ``knowledge`` is already censored;
        when it is

        empty ``refusal`` is a Korean reason.'
...
    Refusal:
      properties:
        code:
          type: string
          enum:
          - no_task
          - blocked_by_policy
          - no_knowledge
          title: Code
        message:
          type: string
          title: Message
      type: object
      required:
      - code
      - message
      title: Refusal
...
    TaskRef:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
      type: object
      required:
      - id
      - name
      title: TaskRef
...
    TeamCreateRequest:
      properties:
        task_id:
          anyOf:
          - type: string
            pattern: ^[a-z][a-z0-9_-]{0,31}$
          - type: 'null'
          title: Task Id
          description: 생략 시 name 에서 slug
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        description:
          type: string
          maxLength: 4000
          minLength: 1
          title: Description
        sandbox:
          anyOf:
          - type: string
          - type: 'null'
          title: Sandbox
          description: 선언된 샌드박스로 격리할 때만. 기본은 기본 샌드박스
      type: object
      required:
      - name
      - description
      title: TeamCreateRequest
      description: '`POST /teams`: requirements in natural language → resident task team (supervisor +
        members).'
    TeamList:
      properties:
        teams:
          items:
            $ref: '#/components/schemas/TeamResponse'
          type: array
          title: Teams
      type: object
      required:
      - teams
      title: TeamList
...
    TeamResponse:
      properties:
        team_id:
          type: string
          title: Team Id
        task:
          $ref: '#/components/schemas/TaskRef'
        pattern:
          $ref: '#/components/schemas/TeamPatternView'
        supervisor:
          type: string
          title: Supervisor
        members:
          items:
            $ref: '#/components/schemas/TeamMemberView'
          type: array
          title: Members
        sandbox:
          anyOf:
          - type: string
          - type: 'null'
          title: Sandbox
        status:
          type: string
          enum:
          - ready
          - applying
          - failed
          - declared
          title: Status
        applied:
          anyOf:
          - $ref: '#/components/schemas/TeamApplied'
          - type: 'null'
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        created_at:
          type: string
          title: Created At
          default: ''
      type: object
      required:
      - team_id
      - task
      - pattern
      - supervisor
      - members
      - sandbox
      - status
      title: TeamResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError